Yanis

+++ Backend/src/controllers/deleteacount.ts
@@ -1,6 +1,7 @@
 import { User, ListeSouhait } from '../models/db.js';
+import { Request, Response } from 'express'
 
-export async function DeleteAccount(req, res) {
+export async function DeleteAccount(req: Request, res: Response) {
   //Suppression de la liste de souhait associé
   const UserListSouhait = await ListeSouhait.findAll({ where: { utilisateur_id: req.params.id } });
   await UserListSouhait.forEach(UserSouhait => {
@@ -15,9 +16,11 @@ export async function DeleteAccount(req, res) {
   const DeleteUser = await User.findOne({
     where: { id: req.params.id }
   })
-  await DeleteUser.destroy({}).then(function (item) {
-    res.send({ "success": "Suppression du compte effecuté" })
-  }).catch(function (err) {
-    res.send({ "error": err });
-  })
+  if (DeleteUser != null) {
+    await DeleteUser.destroy({}).then(function (item) {
+      res.send({ "success": "Suppression du compte effecuté" })
+    }).catch(function (err) {
+      res.send({ "error": err });
+    })
+  }
 }
